% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{copy.data}
\alias{copy.data}
\title{Copy data the clipboard}
\usage{
copy.data()
}
\description{
Method will copy a value of the selected variable into
             the clipboard.
}
\details{
Method will find the first word that would be a valid variable name
         which is 'touched' by the cursor or selection. In the case there is
         multiple cursors/selection only the first one (the primary one) will
         be considered.
}

