% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{copy.to.clipboard}
\alias{copy.to.clipboard}
\title{Write to Clipboard}
\usage{
copy.to.clipboard(tsv = NULL)
}
\arguments{
\item{tsv}{The tab septarated values string}
}
\value{
Methood invisibly returns \code{TRUE} if \code{tsv} was successfully
        copied to the clipboard, or \code{FALSE} otherwise.
}
\description{
If there is something to write to clipboard then write it.
}

