% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.pipe.R
\name{extract.positions}
\alias{extract.positions}
\title{Extract selections}
\usage{
extract.positions()
}
\value{
List of selection ranges.
}
\description{
Method collects all selection ranges in the active documnent.
}

