% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.slash.R
\name{flip}
\alias{flip}
\title{Flips slashes orientation}
\usage{
flip(text = character(0))
}
\arguments{
\item{text}{character where slashes orientation will be flipped}
}
\value{
Method returns string with reversed orientation of slashes.
}
\description{
Method reverse orientation of slashes in the string
}
\details{
If method find forward or backward slash in the \code{text} then
         it will reverse its orientation. In the case there are no slashes
         in the \code{text}, unmodified \code{text} is returned.
}
\examples{
flip('/usr\\\\bin/') == flip(flip(flip('/usr\\\\bin/')))
flip('no change') == 'no change'

}

