% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.slash.R
\name{flip.slash}
\alias{flip.slash}
\title{Flips slashes orientation in code selections or from clipboard}
\usage{

}
\description{
Method reverse orientation of slashes in the selected code
             or in the clipboard.
}
\details{
Method will go through all code selections and reverse orientation
         of all slashes found. In the case there is no content selected or
         only white-space is selected, content of the clipboard will be put
         there - again with all slashes reversed.
}

