% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{get.error.message}
\alias{get.error.message}
\title{Print error message}
\usage{
get.error.message(type = list(type = NULL))
}
\arguments{
\item{type}{The list of 4 values:
\describe{
    \item{name}{name of the variable - not used}
    \item{type}{type of the variable}
    \item{value}{value of the variable - not used}
    \item{supported}{\code{TRUE} if tsv can be generated,
                     \code{FALSE} otherwise - not used}
}}
}
\value{
error message
}
\description{
Method returns appropriate error message.
}

