% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{get.tsv.array}
\alias{get.tsv.array}
\title{Generates tsv for array}
\usage{
get.tsv.array(type = list(name = "", type = "array", value = NULL, supported =
  T))
}
\arguments{
\item{type}{The list of 4 values (as returned by \code{\link{detect.type}}):
\describe{
    \item{name}{name of the variable}
    \item{type}{should be \code{array} - not checked, not used}
    \item{value}{an array}
    \item{supported}{should be \code{TRUE} - not checked, not used}
}}
}
\value{
tsv string
}
\description{
Method generates tsv string for an array
}
\details{
1-dimensional array will be coverted to vector and method
         \code{\link{get.tsv.vector}} will be called.
         2-dimensional array will be coverted to matrix and method
         \code{\link{get.tsv.matrix}} will be called.
         3 and more dimensional array will be flattened to matrix and method
         \code{\link{get.tsv.matrix}} will be called. Matrix will have
         \code{N+1} columns where \code{N} is a number of dimensions and
         \code{M} or \code{M+1} rows, where \code{M} is a product of array
         dimensions. E.g. if array has following dimensions
         \code{dim=c(2, 4, 2)}, then the output table will have
         \code{N=3+1=4} columns and \code{M=2*4*2=16} rows. If array
         dimensions are named then header row will be added. First \code{N}
         columns will be take names from dimensions names and the last column
         will be named after variable. Missing names will stay empty.
}
\seealso{
Other generate.tsv: \code{\link{get.tsv.data.frame}},
  \code{\link{get.tsv.matrix}},
  \code{\link{get.tsv.table}},
  \code{\link{get.tsv.vector}}, \code{\link{get.tsv}}
}

