% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{get.tsv.atomic}
\alias{get.tsv.atomic}
\title{Generate tsv for vector}
\usage{
get.tsv.atomic(type = list(name = NULL, type = "atomic", value = NULL,
  supported = T))
}
\arguments{
\item{type}{The list of 4 values (as returned by \code{\link{detect.type}}):
\describe{
    \item{name}{name of the variable - not used}
    \item{type}{should be \code{atomic} - not checked, not used}
    \item{value}{an atomic vector}
    \item{supported}{should be \code{TRUE} - not checked, not used}
}}
}
\value{
tsv string
}
\description{
Method generates tsv string for an atomic vector
}
\details{
1 or 2 rows seperated with \code{\\n} will be generated. If it is
         a named vector, first row will be a tab separated list of names.
         Second row will be a \code{\\t} separated list of vector values.
}
\seealso{
Other generate.tsv: \code{\link{get.tsv.data.frame}},
  \code{\link{get.tsv.matrix}}, \code{\link{get.tsv}}
}

