% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{get.tsv.data.frame}
\alias{get.tsv.data.frame}
\title{Generates tsv for data frame}
\usage{
get.tsv.data.frame(type = list(name = "", type = "data.frame", value = NULL,
  supported = T))
}
\arguments{
\item{type}{The list of 4 values (as returned by \code{\link{detect.type}}):
\describe{
    \item{name}{name of the variable}
    \item{type}{should be \code{data.frame} - not checked, not used}
    \item{value}{a data frame}
    \item{supported}{should be \code{TRUE} - not checked, not used}
}}
}
\value{
tsv string
}
\description{
Method generates tsv string for a data frame
}
\details{
Method will return N or N+1 rows and M or M+1 tab separated columns.
         +1 in both cases is when row names and/or column names exist.
         If both row and column names exist then top left corner cell will
         contain the \code{name} of the variable.
}
\seealso{
Other generate.tsv: \code{\link{get.tsv.array}},
  \code{\link{get.tsv.matrix}},
  \code{\link{get.tsv.table}},
  \code{\link{get.tsv.vector}}, \code{\link{get.tsv}}
}

