% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{get.tsv.matrix}
\alias{get.tsv.matrix}
\title{Generates tsv for matrix}
\usage{
get.tsv.matrix(type = list(name = "", type = "matrix", value = NULL, supported
  = T))
}
\arguments{
\item{type}{The list of 4 values (as returned by \code{\link{detect.type}}):
\describe{
    \item{name}{name of the variable}
    \item{type}{should be \code{matrix} - not checked, not used}
    \item{value}{matrix}
    \item{supported}{should be \code{TRUE} - not checked, not used}
}}
}
\value{
tsv string
}
\description{
Method generates tsv string for a matrix
}
\details{
Method will return N or N+1 rows and M or M+1 tab separated columns.
         +1 in both cases is when row names and/or column names exist.
         If both row and column names exist then top left corner will be
         constructed from dimension names if they exists in the format:
         Row names dimension name, backslash, column names dimension name.
         In the case there are no dimension names, the \code{name}
         of the variable will be used for the top left corner cell.
}
\seealso{
Other generate.tsv: \code{\link{get.tsv.array}},
  \code{\link{get.tsv.data.frame}},
  \code{\link{get.tsv.table}},
  \code{\link{get.tsv.vector}}, \code{\link{get.tsv}}
}

