% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{get.tsv}
\alias{get.tsv}
\title{Generates tsv}
\usage{
get.tsv(type = list(name = NULL, type = NULL, value = NULL, supported = F))
}
\arguments{
\item{type}{The list of 4 values (as returned by \code{\link{detect.type}}):
\describe{
    \item{name}{name of the variable - not used}
    \item{type}{type of the variable}
    \item{value}{value of the variable}
    \item{supported}{if \code{TRUE} the tsv is generated}
}}
}
\value{
tsv string
}
\description{
Method generates tsv string
}
\details{
if the \code{type} is \code{supported} then it will generate
         appropriate tsv representation of the object.
}
\seealso{
Other generate.tsv: \code{\link{get.tsv.array}},
  \code{\link{get.tsv.data.frame}},
  \code{\link{get.tsv.matrix}},
  \code{\link{get.tsv.table}}, \code{\link{get.tsv.vector}}
}

