% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{get.tsv.table}
\alias{get.tsv.table}
\title{Generates tsv for table}
\usage{
get.tsv.table(type = list(name = "", type = "table", value = NULL, supported =
  T))
}
\arguments{
\item{type}{The list of 4 values (as returned by \code{\link{detect.type}}):
\describe{
    \item{name}{name of the variable}
    \item{type}{should be \code{table} - not checked, not used}
    \item{value}{an table}
    \item{supported}{should be \code{TRUE} - not checked, not used}
}}
}
\value{
tsv string
}
\description{
Method generates tsv string for a table
}
\details{
blank
}
\seealso{
Other generate.tsv: \code{\link{get.tsv.array}},
  \code{\link{get.tsv.data.frame}},
  \code{\link{get.tsv.matrix}},
  \code{\link{get.tsv.vector}}, \code{\link{get.tsv}}
}

