% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{is.namelegal}
\alias{is.namelegal}
\title{Test validity}
\usage{
is.namelegal(line, position = integer(0))
}
\arguments{
\item{line}{A}

\item{position}{A position of the character in the \code{line}}
}
\description{
Method test if the character at the \code{position}
             in the \code{line}, could be a part of the valid variable name.
}

