% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.pipe.R
\name{update.positions}
\alias{update.positions}
\title{Update positions}
\usage{
\method{update}{positions}(positions, index, row, column)
}
\arguments{
\item{positions}{List of documents positions/ranges.}

\item{index}{Index to the \code{list} - element to update}

\item{row}{new last row of the \code{index} selection
(will be incermented by 1)}

\item{column}{new last column of the \code{index} selection
(will be incermented by 1)}
}
\value{
Updated list of positions
}
\description{
Method update selection positions in the document.
}
\details{
The function of the method is to keep track of changes and movement
         of the rows and columns in the documemnt when selections are
         shrinked or expanded.
}

