% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.R
\name{get.tsv.array}
\alias{get.tsv.array}
\title{Generates tsv for array}
\usage{
get.tsv.array(type = list(name = "", type = "array", value = NULL, supported =
  T))
}
\arguments{
\item{type}{The list of 4 values (as returned by \code{\link{detect.type}}):
\describe{
    \item{name}{name of the variable}
    \item{type}{should be \code{array} - not checked, not used}
    \item{value}{an array}
    \item{supported}{should be \code{TRUE} - not checked, not used}
}}
}
\value{
tsv string
}
\description{
Method generates tsv string for a data frame
}
\details{
Method will return ...
}
\seealso{
Other generate.tsv: \code{\link{get.tsv.data.frame}},
  \code{\link{get.tsv.matrix}},
  \code{\link{get.tsv.vector}}, \code{\link{get.tsv}}
}

